refactor http header parsing into HttpHeadersRule, update hyphenate rule to use new class
